# api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Heimdallr API
  version: 1.0.0
  description: API for managing applications, releases, compliance reports, servers, agents, providers, automations, and automation jobs.
servers:
  - url: /api
    description: API base path
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/server:
    get:
      tags:
        - Server
      summary: List servers
      operationId: listServers
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: agent_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
          description: Return only servers where the given agent is installed
      responses:
        "200":
          description: Servers returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Server
      summary: Create server
      operationId: createServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerCreateRequest"
      responses:
        "201":
          description: Server created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/agent:
    get:
      tags:
        - Agent
      summary: List agents
      operationId: listGlobalAgents
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: unassigned
          in: query
          required: false
          schema:
            type: boolean
          description: When true, return only agents not linked to any server
        - name: server_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
          description: Return only agents installed on the given server
        - name: agent_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
          description: Return only the agent with the given ID
      responses:
        "200":
          description: Agents returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Agent
      summary: Register an unassigned agent
      operationId: createUnassignedAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreateRequest"
      responses:
        "201":
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/agent/{agent_id}:
    get:
      tags:
        - Agent
      summary: Get agent by ID
      operationId: getGlobalAgent
      parameters:
        - $ref: "#/components/parameters/AgentIDPath"
      responses:
        "200":
          description: Agent returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetailDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Agent
      summary: Delete an agent and all server links
      operationId: deleteGlobalAgent
      parameters:
        - $ref: "#/components/parameters/AgentIDPath"
      responses:
        "204":
          description: Agent deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/agent/{agent_id}/server:
    get:
      tags:
        - Agent
      summary: List servers where an agent is installed
      operationId: listAgentServers
      parameters:
        - $ref: "#/components/parameters/AgentIDPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Servers returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentServerListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/server/{server_id}/agent:
    get:
      tags:
        - Agent
      summary: List agents installed on a server
      operationId: listAgents
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Agents returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Agent
      summary: Create or link an agent on a server
      operationId: createAgent
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerAgentRequest"
      responses:
        "201":
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/server/{server_id}/agent/{agent_id}:
    get:
      tags:
        - Agent
      summary: Get agent by ID
      operationId: getAgent
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
        - $ref: "#/components/parameters/AgentIDPath"
      responses:
        "200":
          description: Agent returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Agent
      summary: Detach an agent from a server
      operationId: detachAgent
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
        - $ref: "#/components/parameters/AgentIDPath"
      responses:
        "204":
          description: Agent detached successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API token passed as a Bearer token in the Authorization header
    SessionCookie:
      type: apiKey
      in: cookie
      name: heimdallr_session
      description: HttpOnly browser session cookie
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    ServerIDPath:
      name: server_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AgentIDPath:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UUID:
      type: string
      format: uuid
      x-go-type: uuid.UUID
      x-go-type-import:
        path: github.com/google/uuid
        name: uuid
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: invalid request body
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    ServerMetadata:
      type: object
      properties: {}
    ServerRelationSummary:
      type: object
      required:
        - agent_count
        - job_count
        - release_count
      properties:
        agent_count:
          type: integer
          minimum: 0
        job_count:
          type: integer
          minimum: 0
        release_count:
          type: integer
          minimum: 0
    Server:
      type: object
      required:
        - id
        - hostname
        - metadata
        - operating_system
        - hypervisor
        - location
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        hostname:
          type: string
          minLength: 1
          maxLength: 255
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
        operating_system:
          type: string
        hypervisor:
          type: string
        location:
          type: string
    ServerListItem:
      allOf:
        - $ref: "#/components/schemas/Server"
        - type: object
          required:
            - relations
          properties:
            relations:
              $ref: "#/components/schemas/ServerRelationSummary"
...
    ServerCreateRequest:
      type: object
      required:
        - hostname
      properties:
        hostname:
          type: string
          minLength: 1
          maxLength: 255
        operating_system:
          type: string
          maxLength: 255
        hypervisor:
          type: string
          maxLength: 255
        location:
          type: string
          maxLength: 255
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
        agent_ids:
          type: array
          items:
            $ref: "#/components/schemas/UUID"
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentCreateRequest"
...
    ServerDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Server"
...
    ServerListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServerListItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
    Agent:
      type: object
      required:
        - id
        - name
        - type
        - version
        - metadata
        - server_count
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          type: string
        version:
          type: string
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
        server_count:
          type: integer
          minimum: 0
    AgentServer:
      type: object
      required:
        - id
        - hostname
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        hostname:
          type: string
    AgentDetail:
      allOf:
        - $ref: "#/components/schemas/Agent"
        - type: object
          required:
            - servers
          properties:
            servers:
              type: array
              items:
                $ref: "#/components/schemas/AgentServer"
    ServerAgentRequest:
      type: object
      properties:
        agent_id:
          allOf:
            - $ref: "#/components/schemas/UUID"
          nullable: true
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          type: string
          maxLength: 255
        version:
          type: string
          maxLength: 255
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
    AgentCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          type: string
          maxLength: 255
        version:
          type: string
          maxLength: 255
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
    AgentDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Agent"
    AgentDetailDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/AgentDetail"
    AgentServerListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AgentServer"
        pagination:
          $ref: "#/components/schemas/Pagination"
    AgentListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        pagination:
          $ref: "#/components/schemas/Pagination"
...